backend/requirements.txt
# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
sse-starlette==1.8.2

# LangGraph and LangChain
langgraph==0.0.20
langchain==0.1.0
langchain-core==0.1.0

# LLM Providers
google-generativeai==0.3.2
langchain-google-genai==0.0.6
# openai==1.10.0  # Uncomment when adding OpenAI support
# anthropic==0.18.0  # Uncomment when adding Claude support

# Data validation
pydantic==2.6.0
pydantic-settings==2.1.0

# Database and SQL
aiosqlite==0.19.0
sqlparse==0.4.4

# Utilities
python-Levenshtein==0.23.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
httpx==0.26.0
